requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
alembic==1.13.0
pydantic-settings==2.5.0
redis==5.1.0
celery==5.4.0
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.3.0
openai==1.50.0
praw==7.7.0
youtube-transcript-api==0.6.2
python-dotenv==1.0.1
numpy==1.26.4
scikit-learn==1.5.2
torch==2.4.1
torchvision==0.19.1
Pillow==10.4.0
